make sure memcached can connect to servers before performing any operations

+++ src/Adapters/Memcached.php
@@ -15,42 +15,77 @@ class Memcached implements \CacheExchange\Interfaces\DatastoreInterface
   protected $cache;
 
   /**
+   * Can memcached connect to the servers?
+   * @var bool
+   */
+  public $ready = false;
+
+  /**
    * @param $servers
    */
   public function __construct($servers = [])
   {
     $this->cache = new \Memcached();
     $this->cache->addServers($servers);
+
+    if ($this->cache->getStats() !== false) {
+      // can connect to servers
+      $this->ready = true;
+    }
   }
 
   public function set(string $key, mixed $value, int $ttl = 0): bool
   {
+    if (!$this->ready) {
+      return false;
+    }
+
     return $this->cache->set($key, $value, $ttl);
   }
 
   public function get(string $key): mixed
   {
+    if (!$this->ready) {
+      return false;
+    }
+
     return $this->cache->get($key);
   }
 
   public function exists(string $key): bool
   {
+    if (!$this->ready) {
+      return false;
+    }
+
     $data = $this->get($key);
     return !empty($data);
   }
 
   public function delete(string $key): bool
   {
+    if (!$this->ready) {
+      return false;
+    }
+
     return $this->cache->delete($key);
   }
 
   public function clear(): bool
   {
+    if (!$this->ready) {
+      return false;
+    }
+
     return $this->cache->flush();
   }
 
   public function getKeys(): array|false
   {
+    if (!$this->ready) {
+      return false;
+    }
+
     return $this->cache->getAllKeys();
   }
 }
